refactor: replace SidebarSection with Library component and enhance search functionality
